#329 extract LanguageType to separate file
